simplify code structure
