+++ src/ReinforcementLearningCore/src/core/run.jl
@@ -1,3 +1,17 @@
+export AbstractStage,
+    PreExperimentStage,
+    PostExperimentStage,
+    PreEpisodeStage,
+    PostEpisodeStage,
+    PreActStage,
+    PostActStage,
+    PRE_EXPERIMENT_STAGE,
+    POST_EXPERIMENT_STAGE,
+    PRE_EPISODE_STAGE,
+    POST_EPISODE_STAGE,
+    PRE_ACT_STAGE,
+    POST_ACT_STAGE
+
 import Base: run
 
 function run(
@@ -49,3 +63,30 @@ function _run(policy::AbstractPolicy, env::AbstractEnv, stop_condition, hook::Ab
     hook(POST_EXPERIMENT_STAGE, policy, env)
     hook
 end
+
+#####
+# Stage
+#####
+
+abstract type AbstractStage end
+
+struct PreExperimentStage <: AbstractStage end
+const PRE_EXPERIMENT_STAGE = PreExperimentStage()
+
+struct PostExperimentStage <: AbstractStage end
+const POST_EXPERIMENT_STAGE = PostExperimentStage()
+
+struct PreEpisodeStage <: AbstractStage end
+const PRE_EPISODE_STAGE = PreEpisodeStage()
+
+struct PostEpisodeStage <: AbstractStage end
+const POST_EPISODE_STAGE = PostEpisodeStage()
+
+struct PreActStage <: AbstractStage end
+const PRE_ACT_STAGE = PreActStage()
+
+struct PostActStage <: AbstractStage end
+const POST_ACT_STAGE = PostActStage()
+
+(p::AbstractPolicy)(::AbstractStage, ::AbstractEnv) = nothing
+(p::AbstractPolicy)(::AbstractStage, ::AbstractEnv, action) = nothing
