--- src/ReinforcementLearningCore/src/extensions/ReinforcementLearningBase.jl
@@ -1,6 +1,0 @@
-using AbstractTrees
-
-Base.show(io::IO, p::AbstractPolicy) =
-    AbstractTrees.print_tree(io, StructTree(p), maxdepth=get(io, :max_depth, 10))
-
-is_expand(::AbstractEnv) = false
