+++ src/ReinforcementLearningCore/src/extensions/extensions.jl
@@ -2,6 +2,5 @@ include("ArrayInterface.jl")
 include("Flux.jl")
 include("CUDA.jl")
 include("Zygote.jl")
-include("ReinforcementLearningBase.jl")
 include("ElasticArrays.jl")
 include("Distributions.jl")
