--- src/ReinforcementLearningCore/src/policies/agents/agents.jl
@@ -1,3 +1,0 @@
-include("named_policy.jl")
-include("agent.jl")
-include("multi_agent.jl")
