--- src/ReinforcementLearningCore/src/policies/q_based_policies/learners/approximators/approximators.jl
@@ -1,3 +1,0 @@
-include("abstract_approximator.jl")
-include("tabular_approximator.jl")
-include("neural_network_approximator.jl")
