--- src/ReinforcementLearningCore/src/policies/q_based_policies/q_based_policies.jl
@@ -1,3 +1,0 @@
-include("learners/learners.jl")
-include("explorers/explorers.jl")
-include("q_based_policy.jl")
