+++ src/ReinforcementLearningCore/src/utils/utils.jl
@@ -1,5 +1,4 @@
-include("printing.jl")
-include("base.jl")
+include("basic.jl")
 include("device.jl")
-include("sum_tree.jl")
-include("processors.jl")
+include("stack_frames.jl")
+include("explorers/explorers.jl")
